(ASC-1084) Verify galera is up and running in test
This commit to verify that galera is up and running in test level.
There is the same verification of galera in converge tasks, however,
the converge task of galera verification and other verification like
checking nova compute services in converge tasks will be removed when
the healtch checks are implimented and executed before entering system
tests.
